backend/requirements.txt

fastapi
uvicorn[standard]
sqlalchemy
asyncpg
psycopg2-binary
python-dotenv
sqlmodel
python-jose[cryptography]
passlib[bcrypt]
httpx
openai
mcp
alembic
